+++ installer/src/DigitalOcean/DnsValidated.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Cocotte\DigitalOcean;
+
+interface DnsValidated
+{
+
+}